refactor: move arguments builders into dedicated files
